abstract a new ckpt hanlder
